requirements.txt
fastapi
uvicorn
python-telegram-bot
openai
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
pytesseract
Pillow
langchain
langchain-community